move MediaPlatformStream class
