+++ MonoGame.Framework/XNA.Framework.DesktopGL.csproj
@@ -160,6 +160,7 @@
     <Compile Include="Media\DesktopGL\ConcreteMediaFactory.cs" />
     <Compile Include="Media\DesktopGL\ConcreteMediaLibrary.cs" />
     <Compile Include="Media\DesktopGL\ConcreteMediaPlayer.cs" />
+    <Compile Include="Media\DesktopGL\MediaPlatformStream.cs" />
     <Compile Include="Media\DesktopGL\ConcreteSong.cs" />
     <Compile Include="Media\DesktopGL\ConcreteVideoPlayer.cs" />
     <Compile Include="Media\DesktopGL\ConcreteVideo.cs" />
